team_hatakeyama/1_DataPreparation/CommonCrawl/30tokenize/requirements_nedo2024.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.20.3
idna==3.6
mmh3==4.1.0
multidict==6.0.5
multiprocess==0.70.16
#numpy==1.26.4
packaging==23.2
pandas==2.2.1
plac==1.4.3
pyarrow==15.0.0
pyarrow-hotfix==0.6
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
unidic==1.1.0
urllib3==2.2.1
wasabi==0.10.1
xxhash==3.4.1
yarl==1.9.4
################
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105

mwxml==0.3.3
requests==2.31.0
chardet
hojichar==0.9.0
mecab-python3==1.0.8
fugashi[unidic] # v1.3.0 

###############################
ninja==1.11.1
packaging
pip>=23.1
protobuf<3.21.0
pybind11
regex
six
wheel

# Hugging Face
accelerate>=0.23.0
peft>=0.5.0
tokenizers>=0.14.0
transformers>=4.34.0
trl>=0.7.2

# Logging
tensorboard
wandb

# Pre-training & Fine-tuning
bitsandbytes
nltk
sentencepiece

# Tensor operation
einops
numpy
